+++ src/main/java/ru/alex/courseModel/reposttory/CourseRepo.java
@@ -3,5 +3,5 @@ package ru.alex.courseModel.reposttory;
 import org.springframework.data.repository.CrudRepository;
 import ru.alex.courseModel.entity.Course;
 
-public interface CourseRepo extends CrudRepository<Course, Long> {
+public interface CourseRepo extends CrudRepository<Course, Integer> {
 }
